Delete image files after user delete them from website

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product');
-const {validationResult} = require('express-validator')
+const { validationResult } = require('express-validator')
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
   let message = req.flash('error');
@@ -143,7 +144,8 @@ exports.postEditProduct = (req, res, next) => {
     product.title = updatedTitle;
     product.price = updatedPrice;
     product.description = updatedDesc;
-    if(image){  
+    if(image){ 
+      fileHelper.deleteFile(product.imageUrl); 
       product.imageUrl = image.path; 
     }
     return product.save()
@@ -176,13 +178,19 @@ exports.getProducts = (req, res, next ) => {
 
 exports.postDeleteProduct = (req, res, next) => {   
   const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id})
-    .then(() => {
-      res.redirect('/admin/products');
-    })
-    .catch(err =>{
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(error);
-    });
+  Product.findById(prodId).then(product=>{
+    if(!product){
+      return next(new Error('Product not found!'))
+    }
+    fileHelper.deleteFile(product.imageUrl); 
+    return Product.deleteOne({_id: prodId, userId: req.user._id});
+  })
+  .then(() => {
+    res.redirect('/admin/products');
+  })
+  .catch(err =>{
+    const error = new Error(err);
+    error.httpStatusCode = 500;
+    return next(error);
+  });
 };
